requirement.txt
certifi==2024.12.14
charset-normalizer==3.4.1
contourpy==1.3.1
cycler==0.12.1
fonttools==4.55.3
idna==3.10
kiwisolver==1.4.8
matplotlib==3.10.0
numpy==2.2.1
opencv-python==4.10.0.84
packaging==24.2
pillow==11.1.0
pupil_apriltags==1.0.4.post10
pynetworktables==2021.0.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.17.0
urllib3==2.3.0
